backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
python-dotenv==1.0.1
httpx==0.28.1
GitPython==3.1.44
langgraph==0.2.70
google-generativeai==0.8.4
